requirements.txt
openai-whisper
numpy
soundfile